+++ src/components/Configuration/components.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Dropdown from 'react-dropdown';
 
 export const Modes = props => {
   return <a href="#">{props.children}</a>;
